Move scoped_guard.hpp out of include/deepstream
