+++ src/scope_guard.hpp
