+++ test/event.cpp
@@ -23,7 +23,7 @@
 #include <deepstream/event.hpp>
 #include <deepstream/message.hpp>
 #include <deepstream/message_builder.hpp>
-#include <deepstream/scope_guard.hpp>
+#include "../src/scope_guard.hpp"
 
 namespace deepstream {
