+++ test/parser.cpp
@@ -29,7 +29,7 @@
 #include <deepstream/message_builder.hpp>
 #include <deepstream/parser.hpp>
 #include <deepstream/random.hpp>
-#include <deepstream/scope_guard.hpp>
+#include "../src/scope_guard.hpp"
 
 extern "C" {
 #include "lexer.h"
